Migrando las funciones de Typeahead a su propio archivo
Este cambio es un paso más hacia la posibilidad de eliminar ui.js.
