+++ frontend/www/js/omegaup/components/course/ProblemList.vue
@@ -135,7 +135,7 @@
                   >{{ T.wordsProblem }}
                   <omegaup-autocomplete
                     class="form-control"
-                    v-bind:init="el =&gt; UI.problemTypeahead(el)"
+                    v-bind:init="el => typeahead.problemTypeahead(el)"
                     v-model="problemAlias"
                   ></omegaup-autocomplete
                 ></label>
@@ -182,7 +182,7 @@
 import { Vue, Component, Prop, Watch } from 'vue-property-decorator';
 import { omegaup } from '../../omegaup';
 import T from '../../lang';
-import * as UI from '../../ui';
+import * as typeahead from '../../typeahead';
 import Autocomplete from '../Autocomplete.vue';
 
 @Component({
@@ -196,7 +196,7 @@ export default class CourseProblemList extends Vue {
   @Prop() taggedProblems!: omegaup.Problem[];
   @Prop() selectedAssignment!: omegaup.Assignment;
 
-  UI = UI;
+  typeahead = typeahead;
   T = T;
   assignment: Partial<omegaup.Assignment> = {};
   showForm = false;
