+++ frontend/www/js/omegaup/omegaup.js
@@ -1,8 +1,9 @@
 import { Experiments, EventListenerList, OmegaUp } from './omegaup.ts';
 import API from './api.js';
 import * as UI from './ui.js';
+import * as Typeahead from './typeahead';
 import T from './lang';
-export { API, EventListenerList, Experiments, OmegaUp, T, UI };
+export { API, EventListenerList, Experiments, OmegaUp, T, Typeahead, UI };
 
 if (
   document.readyState === 'complete' ||
