+++ frontend/www/js/omegaup/typeahead.d.ts
@@ -1,0 +1,16 @@
+export function groupTypeahead(
+  elem: HTMLElement,
+  cb: (event: HTMLEvent, val: any) => void,
+): void;
+export function problemTypeahead(
+  elem: HTMLElement,
+  cb: (event: HTMLEvent, val: any) => void,
+): void;
+export function schoolTypeahead(
+  elem: any,
+  cb: (event: HTMLEvent, val: any) => void,
+): void;
+export function userTypeahead(
+  elem: HTMLElement,
+  cb: (event: HTMLEvent, val: any) => void,
+): void;
